feat: ajusts and creation of JsonUtils and GlobalExceptionHandler
