+++ order-service/src/main/java/br/com/microservices/orchestrated/orderservice/core/controller/exception/GlobalExceptionHandler.java
@@ -1,0 +1,23 @@
+package br.com.microservices.orchestrated.orderservice.core.controller.exception;
+
+import br.com.microservices.exception.ValidationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
+
+    @ExceptionHandler(ValidationException.class)
+    public ResponseEntity<ProblemDetail> handleValidationException(ValidationException ex) {
+        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(
+                HttpStatus.BAD_REQUEST,
+                ex.getMessage()
+        );
+        problemDetail.setTitle("Validation Error");
+        return ResponseEntity.badRequest().body(problemDetail);
+    }
+}
