+++ payment-service/src/main/java/br/com/microservices/orchestrated/paymentservice/utils/UtilsConfig.java
@@ -1,4 +1,4 @@
-package br.com.microservices.orchestrated.orderservice.config.utils;
+package br.com.microservices.orchestrated.paymentservice.utils;
 
 import br.com.microservices.dto.services.Event;
 import br.com.microservices.utils.JsonUtils;
