+++ saga-utils/build.gradle
@@ -1,13 +1,19 @@
 plugins {
     id 'java'
+    id 'maven-publish'
 }
 
-group = 'io.github.paulovieirajr'
+group = 'br.com.microservices.saga-utils'
 version = '0.0.3'
 
 repositories {
     mavenLocal()
     mavenCentral()
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_21
+    targetCompatibility = JavaVersion.VERSION_21
 }
 
 dependencies {
