+++ saga-utils/src/main/java/br/com/microservices/dto/Order.java
@@ -1,4 +1,4 @@
-package io.github.paulovieirajr.dto;
+package br.com.microservices.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
