+++ saga-utils/src/main/java/br/com/microservices/dto/OrderProduct.java
@@ -1,4 +1,4 @@
-package io.github.paulovieirajr.dto;
+package br.com.microservices.dto;
 
 public record OrderProduct(
         Product product,
