+++ saga-utils/src/main/java/br/com/microservices/dto/orchestrated/History.java
@@ -1,6 +1,6 @@
-package io.github.paulovieirajr.dto.orchestrated;
+package br.com.microservices.dto.orchestrated;
 
-import io.github.paulovieirajr.enums.SagaStatus;
+import br.com.microservices.enums.SagaStatus;
 
 import java.time.LocalDateTime;
