+++ saga-utils/src/main/java/br/com/microservices/enums/SagaStatus.java
@@ -1,0 +1,4 @@
+package br.com.microservices.enums;
+
+public enum SagaStatus {
+}
